mbtool: Use size_t with looping with a size_t index variable

+++ mbtool/packages.c
@@ -193,7 +193,7 @@ int mb_packages_cleanup(struct packages *pkgs)
         return 0;
     }
 
-    for (unsigned int i = 0; i < pkgs->len; ++i) {
+    for (size_t i = 0; i < pkgs->len; ++i) {
         struct package *pkg = pkgs->list[i];
         package_cleanup(pkg);
     }
